backend/requirements.txt
fastapi~=0.112.0
gunicorn~=22.0.0
uvicorn~=0.30.5

pydantic~=2.8.2
pydantic-settings~=2.4.0

sqlalchemy~=2.0.31
aiosqlite~=0.20.0

pyjwt~=2.9.0
bcrypt~=4.2.0

WTForms~=3.1.2